Impove the efficiency of getting labels for entities/properties
Exlcude all information but labels from Wikibase API, which extremely
reduce the size of caching.

The API actually also supports returning up to 50 labels at a time, so
it will be more efficient to do this in batches. However, this means we
need add specific caching schema instead of a universal schema based on
URL right now.
We might need that for if we sample larger dataset.