Use word-level metric to get entity spans
In ThingTalk, we need the span in the original utterance for each
entity and string. QALD-7 provides a keyword field which include
entity names. However, it may not match the label in Wikidata.

We use word-level f1 or jaccard similarity to find the best match among
keywords given the wikidata label.
The similarity is calculated after removing stop words. And if there is
no match (all keywords yield 0 score) at all, we throw an error.