Update the algorithmm to decide the domain of an entity
We check the value of P31 property and find domains that are
included in the top-level domains in the manifest.

If there is none, we iteratively look at the immediate parent
domains (P279) until we find one.
If there is a tie, we use heuristics based on the domain size
and depth in schema.org type hierarchy to decide the best one