Use special token ENTITY for table name and entity types
As we normalize the output, we should always have "entity" as the table name
and entity type unless the domain information is required to parse the query.

When converting from sparql to thingtalk, we simply replace "ENTITY" with "entity"
in the final thingtalk;
To improve generality, when converting from thingtalk to sparql, we replace
special token "ENTITY" to match any non-whitespace characters, so that in case
the model predicts anything other than "entity", we can still get a match.